+++ mvc/src/test/java/nextstep/mvc/handler/mapping/AnnotationHandlerMappingTest.java
@@ -1,7 +1,10 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.handler.mapping;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.handler.HandlerExecution;
+import nextstep.mvc.handler.mapping.AnnotationHandlerMapping;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
